[RSPEED-1567] Improve how RHEL version is determined from system profile (#184)
We ran into some data where operating_system[“name”] exists but both
major and minor keys are missing.

Instead of assuming that everything needed in operating_system is there
if the name field exists, test that we actually got a value for major. If not,
fall back to using os_release to get the major and minor versions.

Lastly, skip instead of failing if we don’t have the version information.

* Add tests for _normalize_stream
* Add tests for rhel_major_minor
* Update test data
   Add scenarios for:
    - missing system_profile_facts
    - missing major/minor but OS name is
    - present but empty operating_system field in system profile

* Do the same thing for relevant app streams

It feels like we should combine this code since it’s easy to forget to
make the same changes is both places.