Combine module and package results in the relevant response (#177)
The end goal is have one entry in the respone per display name. For relevant
results that query inventory, the distinction between package and module isn’t
helpful, so that was removed from the response.

The fields used to compute the hash for the collection object were pared down to
achieve the desired response.

It looks like there may be some inaccurate module/package data that may be causing duplicate items.
That needs to be fixed in the data, not in the application.