Add EUS and E4S dates for RHEL 8.10 and 9.10 (#200)
This seems to be the best solution even though there are not different E4S or EUS dates for the .10 releases.
Adding E4S and EUS dates means the correct end date will be retrieved if the system has
a E4S or EUS entitlement instead of displaying "Unknown" for the end date.