Change how package data are grouped
The goal of using a set initially was to keep the number package items that
needed to be processed to a minimu. But when system_id was added to the mix, it
remove any possibility of deduplicating across systems. Furthermore, this seems
to have introduced a bug that causes inconsistent matching of packages to hosts.

Change the grouping strategy to collect system IDs into groups based on the
package and its major OS version. I think that gets close to the original goal
while still helping improve performance.